+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * [#223](https://github.com/TOMO-CAT/xmake/issues/223): 更新 repo 时使用 git fetch + git reset 的方式, 避免 pull -f 失败
 * [#216](https://github.com/TOMO-CAT/xmake/issues/216): 出错退出时等待后台子进程结束, 支持用户 ctrl + c 主动中断
 * [#234](https://github.com/TOMO-CAT/xmake/issues/234): 支持 ubuntu:latest docker container
+* [#228](https://github.com/TOMO-CAT/xmake/issues/228): local package 本地编译时正确更新所有 xmake repo
 
 ### Bugs 修复